start working on compatibility with LP64
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@959 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ DCM Framework/DCMDataContainer.h
@@ -70,8 +70,8 @@
 - (unsigned char)nextUnsignedChar;
 - (unsigned short)nextUnsignedShort;
 - (short)nextSignedShort;
-- (unsigned long)nextUnsignedLong;
-- (long)nextSignedLong;
+- (unsigned int)nextUnsignedLong;
+- (int)nextSignedLong;
 - (unsigned long long)nextUnsignedLongLong;
 - (long long)nextSignedLongLong;
 - (float)nextFloat;
